src/jobs_spark/notebooks/python/requirements.txt
pandas
minio
pyarrow
psycopg2-binary